// web-react/src/rpc/client.gen.ts
export interface Meta {
  key?: string
  capo?: number
  time?: string
  tempo?: string
  year?: number
  notes?: Array<string>
}